chore: Update backuppc-app.service to include LDAP discovery script
The `backuppc-app.service` file has been modified to include the execution of the `discover-ldap` script as a pre-start step. This script is responsible for discovering the LDAP settings and generating the necessary environment file for NGINX authentication. By including this script, we improve the configuration process for the LDAP service.

```