chore: Update discover-smarthost script to include SMTP_FROM variable
The `discover-smarthost` script has been updated to include the `SMTP_FROM` variable in the generated environment file. This variable is set to the same value as the `SMTP_USER` variable, ensuring that the correct email address is used as the "From" address for outgoing emails. This update improves the configuration process for the SMTP service.